+++ core/dbt/contracts/state.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 from .graph.manifest import WritableManifest
 from typing import Optional
+from dbt.exceptions import IncompatibleSchemaException
 
 
 class PreviousState:
@@ -10,4 +11,8 @@ class PreviousState:
 
         manifest_path = self.path / 'manifest.json'
         if manifest_path.exists() and manifest_path.is_file():
-            self.manifest = WritableManifest.read(str(manifest_path))
+            try:
+                self.manifest = WritableManifest.read(str(manifest_path))
+            except IncompatibleSchemaException as exc:
+                exc.add_filename(str(manifest_path))
+                raise